DePIN-Mega-System/core-systems/phantom-grid-v3/requirements.txt
# ═════════════════════════════════════════════════════════════════════════════
# Phantom Grid v3.0 - Requirements
# Massive Scale Integration
# ═════════════════════════════════════════════════════════════════════════════

# Core
aiohttp>=3.9.0
aiofiles>=23.0.0
aiohttp-socks>=0.8.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0
pynacl>=1.5.0

# Networking
websockets>=12.0
requests>=2.31.0
pyyaml>=6.0.1

# Data
numpy>=1.24.0
msgpack>=1.0.7
protobuf>=4.24.0

# AI/ML
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
huggingface-hub>=0.19.0

# Optional: Petals
git+https://github.com/bigscience-workshop/petals.git

# Optional: libp2p
# libp2p>=0.1.5

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0
tqdm>=4.66.0
psutil>=5.9.0
